replaced Integer attributes by int and replaced Integer ID's by long

+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/RuleRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues;
 import ca.leaguemanagementsystem.model.leagues.Rule;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RuleRepository extends CrudRepository<Rule, Integer> {
+public interface RuleRepository extends CrudRepository<Rule, long> {
 }
