+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/schedules/PlayerTeamAssignmentRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues.schedules;
 import ca.leaguemanagementsystem.model.leagues.schedules.PlayerTeamAssignment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerTeamAssignmentRepository extends CrudRepository<PlayerTeamAssignment, Integer> {
+public interface PlayerTeamAssignmentRepository extends CrudRepository<PlayerTeamAssignment, long> {
 }
