+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/schedules/events/AttendanceRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues.schedules.events;
 import ca.leaguemanagementsystem.model.leagues.schedules.events.Attendance;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AttendanceRepository extends CrudRepository<Attendance, Integer> {
+public interface AttendanceRepository extends CrudRepository<Attendance, long> {
 }
