+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/schedules/events/MVPRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues.schedules.events;
 import ca.leaguemanagementsystem.model.leagues.schedules.events.MVP;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MVPRepository extends CrudRepository<MVP, Integer> {
+public interface MVPRepository extends CrudRepository<MVP, long> {
 }
