+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/leagues/schedules/events/PlayerStatRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.leagues.schedules.events;
 import ca.leaguemanagementsystem.model.leagues.schedules.events.PlayerStat;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerStatRepository extends CrudRepository<PlayerStat, Integer> {
+public interface PlayerStatRepository extends CrudRepository<PlayerStat, long> {
 }
