+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/dao/users/UserRoleRepository.java
@@ -3,5 +3,5 @@ package ca.leaguemanagementsystem.dao.users;
 import ca.leaguemanagementsystem.model.users.UserRole;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRoleRepository extends CrudRepository<UserRole,Integer> {
+public interface UserRoleRepository extends CrudRepository<UserRole,long> {
 }
