+++ LeagueManagementSystem-Backend/src/main/java/ca/leaguemanagementsystem/model/leagues/schedules/events/PlayerStat.java
@@ -38,170 +38,170 @@ public class PlayerStat {
 
     /* attributes */
 
-    private Integer points; // points made during the game
+    private int points; // points made during the game
 
-    public Integer getPoints() {
+    public int getPoints() {
         return points;
     }
 
-    public void setPoints(Integer value) {
+    public void setPoints(int value) {
         this.points = value;
     }
 
 
-    private Integer rebounds;   // rebounds taken during the game
+    private int rebounds;   // rebounds taken during the game
 
-    public Integer getRebounds() {
+    public int getRebounds() {
         return rebounds;
     }
 
-    public void setRebounds(Integer value) {
+    public void setRebounds(int value) {
         this.rebounds = value;
     }
 
 
-    private Integer assists;    // assists made during the game
+    private int assists;    // assists made during the game
 
-    public Integer getAssists() {
+    public int getAssists() {
         return assists;
     }
 
-    public void setAssists(Integer value) {
+    public void setAssists(int value) {
         this.assists = value;
     }
 
 
-    private Integer steals; // steals made during the game
+    private int steals; // steals made during the game
 
-    public Integer getSteals() {
+    public int getSteals() {
         return steals;
     }
 
-    public void setSteals(Integer value) {
+    public void setSteals(int value) {
         this.steals = value;
     }
 
 
-    private Integer blocks; // blocks made during the game
+    private int blocks; // blocks made during the game
 
-    public Integer getBlocks() {
+    public int getBlocks() {
         return blocks;
     }
 
-    public void setBlocks(Integer value) {
+    public void setBlocks(int value) {
         this.blocks = value;
     }
 
 
-    private Integer fieldGoalsMade;  // # of field goals made during the game
+    private int fieldGoalsMade;  // # of field goals made during the game
 
-    public Integer getFieldGoalsAttempted() {
+    public int getFieldGoalsAttempted() {
         return fieldGoalsAttempted;
     }
 
-    public void setFieldGoalsAttempted(Integer fieldGoalsAttempted) {
+    public void setFieldGoalsAttempted(int fieldGoalsAttempted) {
         this.fieldGoalsAttempted = fieldGoalsAttempted;
     }
 
 
-    private Integer fieldGoalsAttempted;  // # of field goals attempted during the game
+    private int fieldGoalsAttempted;  // # of field goals attempted during the game
 
-    public Integer getFieldGoalsMade() {
+    public int getFieldGoalsMade() {
         return fieldGoalsMade;
     }
 
-    public void setFieldGoalsMade(Integer fieldGoalsMade) {
+    public void setFieldGoalsMade(int fieldGoalsMade) {
         this.fieldGoalsMade = fieldGoalsMade;
     }
 
 
-    private Integer threePtsMade;  // # of 3 pointers made during the game
+    private int threePtsMade;  // # of 3 pointers made during the game
 
-    public Integer getThreePtsMade() {
+    public int getThreePtsMade() {
         return threePtsMade;
     }
 
-    public void setThreePtsMade(Integer threePtsMade) {
+    public void setThreePtsMade(int threePtsMade) {
         this.threePtsMade = threePtsMade;
     }
 
 
-    private Integer threePtsAttempted;  // # of 3 pointers attempted during the game
+    private int threePtsAttempted;  // # of 3 pointers attempted during the game
 
-    public Integer getThreePtsAttempted() {
+    public int getThreePtsAttempted() {
         return threePtsAttempted;
     }
 
-    public void setThreePtsAttempted(Integer threePtsAttempted) {
+    public void setThreePtsAttempted(int threePtsAttempted) {
         this.threePtsAttempted = threePtsAttempted;
     }
 
 
-    private Integer freeThrowsMade; // free throws made during the game
+    private int freeThrowsMade; // free throws made during the game
 
-    public Integer getFreeThrowsMade() {
+    public int getFreeThrowsMade() {
         return freeThrowsMade;
     }
 
-    public void setFreeThrowsMade(Integer freeThrowsMade) {
+    public void setFreeThrowsMade(int freeThrowsMade) {
         this.freeThrowsMade = freeThrowsMade;
     }
 
 
-    private Integer freeThrowsAttempted; // free throws attempted during the game
+    private int freeThrowsAttempted; // free throws attempted during the game
 
-    public Integer getFreeThrowsAttempted() {
+    public int getFreeThrowsAttempted() {
         return freeThrowsAttempted;
     }
 
-    public void setFreeThrowsAttempted(Integer freeThrowsAttempted) {
+    public void setFreeThrowsAttempted(int freeThrowsAttempted) {
         this.freeThrowsAttempted = freeThrowsAttempted;
     }
 
 
-    private Integer turnovers;  // number of turnovers made during the game
+    private int turnovers;  // number of turnovers made during the game
 
-    public Integer getTurnovers() {
+    public int getTurnovers() {
         return turnovers;
     }
 
-    public void setTurnovers(Integer turnovers) {
+    public void setTurnovers(int turnovers) {
         this.turnovers = turnovers;
     }
 
 
-    private Integer fouls;  // number of fouls obtained during the game
+    private int fouls;  // number of fouls obtained during the game
 
-    public Integer getFouls() {
+    public int getFouls() {
         return fouls;
     }
 
-    public void setFouls(Integer fouls) {
+    public void setFouls(int fouls) {
         this.fouls = fouls;
     }
 
 
-    private Integer technicalFouls; // number of technical fouls obtained during the game
+    private int technicalFouls; // number of technical fouls obtained during the game
 
-    public Integer getTechnicalFouls() {
+    public int getTechnicalFouls() {
         return technicalFouls;
     }
 
-    public void setTechnicalFouls(Integer technicalFouls) {
+    public void setTechnicalFouls(int technicalFouls) {
         this.technicalFouls = technicalFouls;
     }
 
 
 
-    private Integer playerStatId;
+    private long playerStatId;
 
     @Id
     @GeneratedValue
-    public Integer getPlayerStatId() {
+    public long getPlayerStatId() {
         return playerStatId;
     }
 
-    public void setPlayerStatId(Integer playerStatId) {
+    public void setPlayerStatId(long playerStatId) {
         this.playerStatId = playerStatId;
     }
 }
